% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics_project.R
\name{add_experiment}
\alias{add_experiment}
\title{Add an experiment to a project}
\usage{
add_experiment(project, name, input)
}
\arguments{
\item{project}{An \code{omics_project}.}

\item{name}{Experiment tag (e.g. \code{"proteomics"}, \code{"rnaseq"}). Must be
unique within the project.}

\item{input}{An \code{omics_input} to attach.}
}
\value{
The modified \code{omics_project}.
}
\description{
Add an experiment to a project
}
\seealso{
Other omics_project: 
\code{\link{experiment_tags}()},
\code{\link{is_omics_project}()},
\code{\link{omics_project}()},
\code{\link{remove_experiment}()}
}
\concept{omics_project}
