% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-import-report.R
\name{add_import_warning}
\alias{add_import_warning}
\title{Append a warning to an \code{ImportReport}}
\usage{
add_import_warning(report, msg)
}
\arguments{
\item{report}{An \code{ImportReport}.}

\item{msg}{Character scalar warning message.}
}
\value{
The updated \code{ImportReport}.
}
\description{
Append a warning to an \code{ImportReport}
}
\keyword{internal}
