% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result-schema.R
\name{check_enrich_result_schema}
\alias{check_enrich_result_schema}
\title{Validate the standard enrichment result schema}
\usage{
check_enrich_result_schema(result_df)
}
\arguments{
\item{result_df}{Data frame to validate.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Validate the standard enrichment result schema
}
\keyword{internal}
