% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{check_feature_matches_expr}
\alias{check_feature_matches_expr}
\title{Check that feature metadata matches an expression matrix}
\usage{
check_feature_matches_expr(expr_mat, feature_df, feature_id_col = "feature_id")
}
\arguments{
\item{expr_mat}{Expression matrix.}

\item{feature_df}{Feature metadata.}

\item{feature_id_col}{Feature identifier column in \code{feature_df}.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Check that feature metadata matches an expression matrix
}
\keyword{internal}
