% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{check_meta_matches_expr}
\alias{check_meta_matches_expr}
\title{Check that sample metadata matches an expression matrix}
\usage{
check_meta_matches_expr(expr_mat, meta_df)
}
\arguments{
\item{expr_mat}{Expression matrix.}

\item{meta_df}{Sample metadata.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Check that sample metadata matches an expression matrix
}
\keyword{internal}
