% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result-schema.R
\docType{data}
\name{DIFF_RESULT_REQUIRED_COLS}
\alias{DIFF_RESULT_REQUIRED_COLS}
\title{Standardized differential analysis result columns}
\format{
An object of class \code{character} of length 17.
}
\usage{
DIFF_RESULT_REQUIRED_COLS
}
\description{
All differential analysis backends (DESeq2, edgeR, limma, t-test, lm)
return a \code{data.frame} with these columns so that downstream visualization
and integration functions can operate uniformly.
}
\keyword{internal}
