% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result-schema.R
\docType{data}
\name{ENRICH_RESULT_REQUIRED_COLS}
\alias{ENRICH_RESULT_REQUIRED_COLS}
\title{Standardized enrichment result columns}
\format{
An object of class \code{character} of length 16.
}
\usage{
ENRICH_RESULT_REQUIRED_COLS
}
\description{
All enrichment backends (ORA, GSEA, GSVA) return a \code{data.frame} with
these columns.
}
\keyword{internal}
