% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics_project.R
\name{experiment_tags}
\alias{experiment_tags}
\title{List experiment tags in a project}
\usage{
experiment_tags(project)
}
\arguments{
\item{project}{An \code{omics_project}.}
}
\value{
Character vector of experiment tags.
}
\description{
List experiment tags in a project
}
\seealso{
Other omics_project: 
\code{\link{add_experiment}()},
\code{\link{is_omics_project}()},
\code{\link{omics_project}()},
\code{\link{remove_experiment}()}
}
\concept{omics_project}
