% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-import-report.R
\docType{data}
\name{IMPORT_REPORT_ROLES}
\alias{IMPORT_REPORT_ROLES}
\title{Sheet role labels recognised by \code{\link[=read_omics]{read_omics()}} and the classifier}
\format{
An object of class \code{character} of length 4.
}
\usage{
IMPORT_REPORT_ROLES
}
\description{
Sheet role labels recognised by \code{\link[=read_omics]{read_omics()}} and the classifier
}
\keyword{internal}
