% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-analysis_bundle.R
\name{is_analysis_bundle}
\alias{is_analysis_bundle}
\title{Test whether an object is an analysis bundle}
\usage{
is_analysis_bundle(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is an analysis bundle
}
