% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-import-report.R
\name{is_import_report}
\alias{is_import_report}
\title{Test whether an object is an \code{ImportReport}}
\usage{
is_import_report(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is an \code{ImportReport}
}
\seealso{
Other io: 
\code{\link{classify_sheet_role}()},
\code{\link{detect_id_columns}()},
\code{\link{detect_orientation}()},
\code{\link{import_report_sheets}()},
\code{\link{import_report_warnings}()},
\code{\link{new_import_report}()},
\code{\link{read_omics}()},
\code{\link{read_quant_files}()}
}
\concept{io}
