% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics_project.R
\name{is_omics_project}
\alias{is_omics_project}
\title{Test whether an object is an \code{omics_project}}
\usage{
is_omics_project(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is an \code{omics_project}
}
\seealso{
Other omics_project: 
\code{\link{add_experiment}()},
\code{\link{experiment_tags}()},
\code{\link{omics_project}()},
\code{\link{remove_experiment}()}
}
\concept{omics_project}
