% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact-registry.R
\name{merge_artifact_registries}
\alias{merge_artifact_registries}
\title{Merge artifact registries}
\usage{
merge_artifact_registries(..., deduplicate = TRUE)
}
\arguments{
\item{...}{Artifact registry objects.}

\item{deduplicate}{Whether to remove duplicated rows after merging.}
}
\value{
Merged artifact registry.
}
\description{
Merge artifact registries
}
\keyword{internal}
