% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result-schema.R
\name{new_diff_result_template}
\alias{new_diff_result_template}
\title{Create an empty diff result template}
\usage{
new_diff_result_template()
}
\value{
Empty \code{data.frame} following the standardized diff-result schema.
}
\description{
Create an empty diff result template
}
\keyword{internal}
