% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result-schema.R
\name{new_enrich_result_template}
\alias{new_enrich_result_template}
\title{Create an empty enrichment result template}
\usage{
new_enrich_result_template()
}
\value{
Empty \code{data.frame} following the standardized enrichment schema.
}
\description{
Create an empty enrichment result template
}
\keyword{internal}
