% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result-schema.R
\name{new_integration_result_template}
\alias{new_integration_result_template}
\title{Create an empty integration result template}
\usage{
new_integration_result_template()
}
\value{
Empty \code{data.frame} following the standardized integration schema.
}
\description{
Create an empty integration result template
}
\keyword{internal}
