% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-analysis_bundle.R
\name{print.analysis_bundle}
\alias{print.analysis_bundle}
\title{Print method for \code{analysis_bundle}}
\usage{
\method{print}{analysis_bundle}(x, ...)
}
\arguments{
\item{x}{An \code{analysis_bundle} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{analysis_bundle}
}
