% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-import-report.R
\name{print.ImportReport}
\alias{print.ImportReport}
\title{Print method for \code{ImportReport}}
\usage{
\method{print}{ImportReport}(x, ...)
}
\arguments{
\item{x}{An \code{ImportReport}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{ImportReport}
}
