% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics_input.R
\name{print.omics_input}
\alias{print.omics_input}
\title{Print method for \code{omics_input}}
\usage{
\method{print}{omics_input}(x, ...)
}
\arguments{
\item{x}{An \code{omics_input} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{omics_input}
}
