% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics_project.R
\name{print.omics_project}
\alias{print.omics_project}
\title{Print method for \code{omics_project}}
\usage{
\method{print}{omics_project}(x, ...)
}
\arguments{
\item{x}{An \code{omics_project} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{omics_project}
}
