% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics_project.R
\name{remove_experiment}
\alias{remove_experiment}
\title{Remove an experiment from a project}
\usage{
remove_experiment(project, name)
}
\arguments{
\item{project}{An \code{omics_project}.}

\item{name}{Experiment tag to remove.}
}
\value{
The modified \code{omics_project}.
}
\description{
Remove an experiment from a project
}
\seealso{
Other omics_project: 
\code{\link{add_experiment}()},
\code{\link{experiment_tags}()},
\code{\link{is_omics_project}()},
\code{\link{omics_project}()}
}
\concept{omics_project}
