% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-standardize.R
\name{standardize_deseq2_continuous_results}
\alias{standardize_deseq2_continuous_results}
\title{Standardize DESeq2 continuous results}
\usage{
standardize_deseq2_continuous_results(
  raw_df,
  feature_df,
  comparison,
  omics_type = "rnaseq"
)
}
\arguments{
\item{raw_df}{Raw DESeq2 result with \code{feature_id}, \code{log2FoldChange},
\code{pvalue}, \code{padj}.}

\item{feature_df}{Feature metadata.}

\item{comparison}{Continuous variable name.}

\item{omics_type}{Omics modality label.}
}
\value{
Standardized diff result data frame.
}
\description{
Standardize DESeq2 continuous results
}
\keyword{internal}
