% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-standardize.R
\name{standardize_edger_group_results}
\alias{standardize_edger_group_results}
\title{Standardize edgeR two-group results}
\usage{
standardize_edger_group_results(
  raw_df,
  feature_df,
  comparison,
  omics_type = "rnaseq"
)
}
\arguments{
\item{raw_df}{Raw edgeR topTags result with \code{feature_id}, \code{logFC}, \code{PValue},
\code{FDR}.}

\item{feature_df}{Feature metadata.}

\item{comparison}{Comparison label.}

\item{omics_type}{Omics modality label.}
}
\value{
Standardized diff result data frame.
}
\description{
Standardize edgeR two-group results
}
\keyword{internal}
