% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-standardize.R
\name{standardize_limma_continuous_results}
\alias{standardize_limma_continuous_results}
\title{Standardize limma continuous results}
\usage{
standardize_limma_continuous_results(
  raw_df,
  feature_df,
  comparison,
  omics_type = "proteomics"
)
}
\arguments{
\item{raw_df}{Raw limma topTable result with \code{feature_id}, \code{P.Value},
\code{adj.P.Val}, \code{spearman_rho}, \code{adj_r_squared}.}

\item{feature_df}{Feature metadata.}

\item{comparison}{Continuous variable name.}

\item{omics_type}{Omics modality label.}
}
\value{
Standardized diff result data frame.
}
\description{
Standardize limma continuous results
}
\keyword{internal}
