% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-standardize.R
\name{standardize_limma_group_results}
\alias{standardize_limma_group_results}
\title{Standardize limma two-group results}
\usage{
standardize_limma_group_results(
  raw_df,
  feature_df,
  comparison,
  omics_type = "proteomics"
)
}
\arguments{
\item{raw_df}{Raw limma topTable result with \code{feature_id}, \code{logFC},
\code{P.Value}, \code{adj.P.Val}.}

\item{feature_df}{Feature metadata.}

\item{comparison}{Comparison label.}

\item{omics_type}{Omics modality label.}
}
\value{
Standardized diff result data frame.
}
\description{
Standardize limma two-group results
}
\keyword{internal}
