% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_input-ops.R
\name{subset_omics_features}
\alias{subset_omics_features}
\title{Subset an \code{omics_input} by feature IDs}
\usage{
subset_omics_features(omics_input, feature_ids)
}
\arguments{
\item{omics_input}{A validated \code{omics_input}.}

\item{feature_ids}{Feature identifiers to retain.}
}
\value{
A new \code{omics_input} containing the selected features.
}
\description{
Subset an \code{omics_input} by feature IDs
}
\seealso{
Other omics_input:
\code{\link{LOG_SCALE_ASSAY_TYPES}},
\code{\link{SUPPORTED_ASSAY_TYPES}},
\code{\link[=check_assay_scale]{check_assay_scale()}},
\code{\link[=drop_meta_na]{drop_meta_na()}},
\code{\link[=infer_assay_type]{infer_assay_type()}},
\code{\link[=is_omics_input]{is_omics_input()}},
\code{\link[=normalize_omics]{normalize_omics()}},
\code{\link[=omics_input]{omics_input()}},
\code{\link[=select_complete_cases]{select_complete_cases()}},
\code{\link[=set_study_design]{set_study_design()}},
\code{\link[=study_design]{study_design()}},
\code{\link[=subset_omics]{subset_omics()}},
\code{\link[=subset_omics_samples]{subset_omics_samples()}},
\code{\link[=summarize_omics]{summarize_omics()}},
\code{\link[=validate_omics_input]{validate_omics_input()}}
}
\concept{omics_input}
