% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-import-report.R
\name{import_report_sheets}
\alias{import_report_sheets}
\title{Access the sheets table from an \code{ImportReport}}
\usage{
import_report_sheets(report)
}
\arguments{
\item{report}{An \code{ImportReport}.}
}
\value{
A \code{data.frame} of detected sheets.
}
\description{
Access the sheets table from an \code{ImportReport}
}
\seealso{
Other io: 
\code{\link{classify_sheet_role}()},
\code{\link{detect_id_columns}()},
\code{\link{detect_orientation}()},
\code{\link{import_report_warnings}()},
\code{\link{is_import_report}()},
\code{\link{new_import_report}()},
\code{\link{read_omics}()},
\code{\link{read_quant_files}()}
}
\concept{io}
