% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-import-report.R
\name{import_report_warnings}
\alias{import_report_warnings}
\title{Access the warnings vector from an \code{ImportReport}}
\usage{
import_report_warnings(report)
}
\arguments{
\item{report}{An \code{ImportReport}.}
}
\value{
Character vector of warnings.
}
\description{
Access the warnings vector from an \code{ImportReport}
}
\seealso{
Other io: 
\code{\link{classify_sheet_role}()},
\code{\link{detect_id_columns}()},
\code{\link{detect_orientation}()},
\code{\link{import_report_sheets}()},
\code{\link{is_import_report}()},
\code{\link{new_import_report}()},
\code{\link{read_omics}()},
\code{\link{read_quant_files}()},
\code{\link{read_summarized_experiment}()}
}
\concept{io}
