% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics_input.R
\name{is_omics_input}
\alias{is_omics_input}
\title{Test whether an object is an \code{omics_input}}
\usage{
is_omics_input(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is an \code{omics_input}
}
\seealso{
Other omics_input: 
\code{\link{drop_meta_na}()},
\code{\link{omics_input}()},
\code{\link{select_complete_cases}()},
\code{\link{subset_omics}()},
\code{\link{subset_omics_features}()},
\code{\link{subset_omics_samples}()},
\code{\link{summarize_omics}()},
\code{\link{validate_omics_input}()}
}
\concept{omics_input}
