% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LOG_SCALE_ASSAY_TYPES}
\alias{LOG_SCALE_ASSAY_TYPES}
\title{Assay types that already sit on a log-like scale}
\format{
An object of class \code{character} of length 5.
}
\usage{
LOG_SCALE_ASSAY_TYPES
}
\description{
Used by \code{\link[=check_assay_scale]{check_assay_scale()}} to decide which direction a scale mismatch
points in, and by callers deciding whether a layer still needs
\code{\link[=normalize_omics]{normalize_omics()}}. Everything else in \link{SUPPORTED_ASSAY_TYPES} is linear.
}
\seealso{
Other omics_input: 
\code{\link{SUPPORTED_ASSAY_TYPES}},
\code{\link{check_assay_scale}()},
\code{\link{drop_meta_na}()},
\code{\link{infer_assay_type}()},
\code{\link{is_omics_input}()},
\code{\link{normalize_omics}()},
\code{\link{omics_input}()},
\code{\link{select_complete_cases}()},
\code{\link{subset_omics}()},
\code{\link{subset_omics_features}()},
\code{\link{subset_omics_samples}()},
\code{\link{summarize_omics}()},
\code{\link{validate_omics_input}()}
}
\concept{omics_input}
\keyword{datasets}
