% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{omicsCore-constants}
\alias{omicsCore-constants}
\alias{SUPPORTED_OMICS_TYPES}
\alias{SUPPORTED_DIFF_ANALYSIS_TYPES}
\alias{SUPPORTED_PREFERENCE}
\alias{SUPPORTED_ENRICH_PREFERENCE}
\alias{DEFAULT_CACHE_MAX_AGE_HOURS}
\alias{CANONICAL_DATABASES}
\title{Package constants}
\format{
An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 6.
}
\usage{
SUPPORTED_OMICS_TYPES

SUPPORTED_DIFF_ANALYSIS_TYPES

SUPPORTED_PREFERENCE

SUPPORTED_ENRICH_PREFERENCE

DEFAULT_CACHE_MAX_AGE_HOURS

CANONICAL_DATABASES
}
\description{
Internal vocabulary used across \code{omicsCore}. Not exported.
}
\keyword{datasets}
\keyword{internal}
