% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-enrichment.R
\name{plot_gsea}
\alias{plot_gsea}
\title{GSEA running-score plot}
\usage{
plot_gsea(bundle, pathway_id, database = NULL)
}
\arguments{
\item{bundle}{An \code{\link[=is_analysis_bundle]{analysis_bundle}} produced by
\code{\link[=run_enrichment]{run_enrichment()}} with \code{type = "gsea"}.}

\item{pathway_id}{Pathway ID (matches \code{pathway_id} in the standardized
table) or pathway name.}

\item{database}{Database key. Required when the bundle holds multiple
databases; ignored when there's only one.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the \code{clusterProfiler::gseaplot2()} running-score curve for a single
pathway in a GSEA enrichment bundle. Requires the \code{enrichplot}
Bioconductor package.
}
\seealso{
Other enrich:
\code{\link[=compare_enrichment]{compare_enrichment()}},
\code{\link[=enrichment_species]{enrichment_species()}},
\code{\link[=filter_enrich_results]{filter_enrich_results()}},
\code{\link[=geneset_cache_status]{geneset_cache_status()}},
\code{\link[=hgnc_map_provenance]{hgnc_map_provenance()}},
\code{\link[=list_gene_sets]{list_gene_sets()}},
\code{\link[=map_ensembl_symbols]{map_ensembl_symbols()}},
\code{\link[=plot_enrichment]{plot_enrichment()}},
\code{\link[=plot_enrichment_comparison]{plot_enrichment_comparison()}},
\code{\link[=plot_gsva_heatmap]{plot_gsva_heatmap()}},
\code{\link[=refresh_geneset_cache]{refresh_geneset_cache()}},
\code{\link[=run_enrichment]{run_enrichment()}},
\code{\link[=run_gsva]{run_gsva()}}
}
\concept{enrich}
